Combine util functions for jit. (#2430)
The JIT compiler used for computing fixed column and the new code for
witgen used two different "util" code snippets. Since the field
implementation for the new witgen code is more advanced, it is now moved
to the generic jit compiler and used by both. It also adds some types
that will later be used by prover functions (PilVec, Callable, etc).
Those types were present already in the fixed column jit code.
