+++ jit-compiler/src/compiler.rs
@@ -13,20 +13,16 @@ use powdr_ast::{
         types::{FunctionType, Type, TypeScheme},
     },
 };
-use powdr_number::{FieldElement, LargeInt};
+use powdr_number::FieldElement;
 
-use crate::{codegen::escape_symbol, CompiledPIL, FixedColFunction};
+use crate::{codegen::escape_symbol, util_code::util_code, CompiledPIL, FixedColFunction};
 
 pub fn generate_glue_code<T: FieldElement>(
     symbols: &[(&str, String)],
     analyzed: &Analyzed<T>,
 ) -> Result<String, String> {
-    if T::BITS > 64 {
-        return Err(format!(
-            "Fields with more than 64 bits not supported, requested {}",
-            T::BITS,
-        ));
-    }
+    let utils = util_code::<T>()?;
+
     let mut glue = String::new();
     let int_int_fun: TypeScheme = Type::Function(FunctionType {
         params: vec![Type::Int],
@@ -54,128 +50,7 @@ pub fn generate_glue_code<T: FieldElement>(
         ));
     }
 
-    Ok(format!(
-        "{PREAMBLE}\n{}\n{glue}\n",
-        field_specific_preamble::<T>()
-    ))
-}
-
-const PREAMBLE: &str = r#"
-#![allow(unused_parens, unused_variables)]
-
-static DEGREE: std::sync::RwLock<Option<ibig::IBig>> = std::sync::RwLock::new(None);
-
-#[no_mangle]
-pub extern "C" fn __set_degree(degree: u64) {
-    *DEGREE.write().unwrap() = Some(ibig::IBig::from(degree));
-}
-
-#[derive(Clone, Copy)]
-struct FieldElement(u64);
-impl From<FieldElement> for u64 {
-    fn from(x: FieldElement) -> u64 {
-        x.0
-    }
-}
-impl From<ibig::IBig> for FieldElement {
-    fn from(x: ibig::IBig) -> Self {
-        FieldElement(u64::try_from(x).unwrap())
-    }
-}
-impl From<FieldElement> for ibig::IBig {
-    fn from(x: FieldElement) -> Self {
-        ibig::IBig::from(x.0)
-    }
-}
-
-#[derive(Clone)]
-enum Callable<Args, Ret> {
-    Fn(fn(Args) -> Ret),
-    Closure(std::sync::Arc<dyn Fn(Args) -> Ret + Send + Sync>),
-}
-impl<Args, Ret> Callable<Args, Ret> {
-    #[inline(always)]
-    fn call(&self, args: Args) -> Ret {
-        match self {
-            Callable::Fn(f) => f(args),
-            Callable::Closure(f) => f(args),
-        }
-    }
-}
-
-#[derive(Clone)]
-struct PilVec<T>(std::sync::Arc<Vec<T>>);
-
-impl<T> PilVec<T> {
-    fn len(&self) -> usize {
-        self.0.len()
-    }
-}
-impl<T> From<Vec<T>> for PilVec<T> {
-    fn from(v: Vec<T>) -> Self {
-        PilVec(std::sync::Arc::new(v))
-    }
-}
-impl<T> std::ops::Index<usize> for PilVec<T> {
-    type Output = T;
-
-    #[inline]
-    fn index(&self, index: usize) -> &T {
-        &self.0[index]
-    }
-}
-    
-
-trait Add {
-    fn add(a: Self, b: Self) -> Self;
-}
-
-impl Add for ibig::IBig {
-    fn add(a: Self, b: Self) -> Self { a + b }
-}
-
-impl<T: Clone> Add for PilVec<T> {
-    fn add(a: Self, b: Self) -> Self {
-        // TODO for a regular "push" or array::map this is very slow.
-        // We could optimize this by sharing a larger backing vector
-        // across prefix instances, allowing to extend the backing vector if
-        // our view is the full vector.
-        PilVec(std::sync::Arc::new(
-            a.0.as_ref().iter().chain(b.0.as_ref()).cloned().collect::<Vec<_>>()
-        ))
-    }
-}
-
-trait FromLiteral {
-    fn from_u64(x: u64) -> Self;
-}
-impl FromLiteral for ibig::IBig {
-    fn from_u64(x: u64) -> Self { ibig::IBig::from(x) }
-}
-impl FromLiteral for FieldElement {
-    fn from_u64(x: u64) -> Self { FieldElement::from(x) }
-}
-
-"#;
-
-fn field_specific_preamble<T: FieldElement>() -> String {
-    let modulus = u64::try_from(T::modulus().to_arbitrary_integer()).unwrap();
-    format!(
-        r#"
-        impl From<u64> for FieldElement {{
-            fn from(x: u64) -> Self {{
-                // TODO this is inefficient.
-                FieldElement(x % {modulus}_u64)
-            }}
-        }}
-        impl Add for FieldElement {{
-            fn add(a: Self, b: Self) -> Self {{
-                // TODO this is inefficient.
-                Self(u64::try_from((u128::from(a.0) + u128::from(b.0)) % u128::from({modulus}_u64)).unwrap())
-            }}
-        }}
-        "#
-    )
+    Ok(format!("{utils}\n{glue}\n",))
 }
 
 const CARGO_TOML: &str = r#"
