+++ jit-compiler/src/includes/field_generic_up_to_64.rs
@@ -126,3 +126,14 @@ impl std::ops::BitOr<FieldElement> for FieldElement {
         Self(self.0 | b.0)
     }
 }
+
+impl From<ibig::IBig> for FieldElement {
+    fn from(x: ibig::IBig) -> Self {
+        FieldElement(u64::try_from(x).unwrap())
+    }
+}
+impl From<FieldElement> for ibig::IBig {
+    fn from(x: FieldElement) -> Self {
+        ibig::IBig::from(IntType::from(x))
+    }
+}
